refactor(unauth_resource_section): move get evaluation data to AddEvaluation component
