+++ src/components/resource-section/resource-section/ResourceSection.js
@@ -15,13 +15,6 @@ const ResourceSection = ({
   } = useGet(endpoints('resource', resourceId));
 
   const {
-    data,
-    isLoading: isLoadingEvaluation,
-    isError: isErrorEvaluation,
-    mutate: updateResourceEvaluation,
-  } = useGet(endpoints('resourceEvaluation', resourceId));
-
-  const {
     data: average_evaluation,
     isLoading: isLoadingAverage,
     isError: isErrorAverage,
@@ -62,7 +55,6 @@ const ResourceSection = ({
     isLoadingResource ||
     isLoadingAverage ||
     isLoadingEvaluations ||
-    isLoadingEvaluation ||
     isLoadingCompleted
   )
     return 'loading';
@@ -71,38 +63,9 @@ const ResourceSection = ({
     isErrorResource ||
     isErrorAverage ||
     isErrorEvaluations ||
-    isErrorEvaluation ||
     isErrorCompleted
   )
     return 'error';
-
-  const showSuccess = () =>
-    toast.current.show({
-      severity: 'success',
-      summary: 'Tu evaluación quedó registrada',
-      detail: 'Gracias por contribuir!',
-    });
-
-  async function handleSubmitForm(evaluation, comment) {
-    if (evaluation < 1) {
-      return;
-    }
-    const requestOptions = {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ evaluation: evaluation, comment: comment }),
-    };
-    const response = await fetch(
-      endpoints('resourceEvaluation', resourceId),
-      requestOptions
-    );
-    await response.json();
-    updateEvaluations();
-    updateAverage();
-    updateResourceEvaluation();
-    onEvaluationSubmitionHandler();
-    showSuccess();
-  }
 
   const resource = {
     name: resourceData.name,
@@ -111,19 +74,18 @@ const ResourceSection = ({
     completed: isCompleted.completed,
   };
 
-  const formOptions = {
-    evaluation: data.evaluation,
-    comment: data.comment,
-    evaluated: data.evaluation ? true : false,
-    handleSubmitForm: handleSubmitForm,
-    toast: toast,
+  const updates = {
+    onEvaluationSubmitionHandler,
+    updateAverage,
+    updateEvaluations,
+    toast,
   };
-
   return (
     <ResourceSidebar
       onHideHandler={() => onHideHandler()}
+      resourceId={resourceId}
       activeResource={resource}
-      formOptions={formOptions}
+      updates={updates}
       evaluations={evaluations}
       checkboxChangeHandler={checkboxChangeHandler}
     />
